feat: monorepo phase 0+1 -- uv workspace root + drift-config capability package (ADR-100)
(cherry picked from commit d9d5315f3eeb5917998638df14710246d7524cce)
